Commit 7 - Proyecto Final de c sharp - Correcciones varias ara poder generar la conexion a la base de datos sin tener hardcodeado el enlace en la logica, y la creacion de un usuario Admin por defecto

+++ SistemaGestionData/DataAccess/UsuariosDataAccess.cs
@@ -56,11 +56,14 @@ public class UsuariosDataAccess
 
     public async Task DeleteUsuario(int id)
     {
-        var usuarioToDelete = await GetOneUsuario(id);
-        if (usuarioToDelete != null)
+        if(id != 1)
         {
-            _context.Usuarios.Remove(usuarioToDelete);
-            await _context.SaveChangesAsync();
+            var usuarioToDelete = await GetOneUsuario(id);
+            if (usuarioToDelete != null)
+            {
+                _context.Usuarios.Remove(usuarioToDelete);
+                await _context.SaveChangesAsync();
+            }
         }
     }
 }
